% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_am_emily.R
\name{i_am_emily}
\alias{i_am_emily}
\title{I Am Emily Function}
\usage{
i_am_emily(name)
}
\arguments{
\item{name}{Any name}
}
\description{
This function makes you feel as special or unspecial as your name.
}
\examples{
i_am_emily("Emily")
i_am_emily("Bob")
}
\keyword{emily}

